[BE-3183] - Adds ability to overwrite metadata
When a :metadata field is passed into the options optional argument of generate-presigned-url, the generated url will have the specified metadata field.

The use case for this is changing how a piece of content is displayed when a user clicks on the presigned url.  BE-3183 shows the :content-disposition field can be used to ensure a document is read as an attachment that downloads to a user's computer as well as changing the filename to ensure that the correct filetype and filename is being downloaded.